FEAT: Rewrite EmailReceiver
