+++ EasyContact.app/server/src/main/java/app/easycontact/server/imap/EmailReceiver.java
@@ -1,138 +1,14 @@
 package app.easycontact.server.imap;
 
-import javax.mail.*;
-import java.util.Properties;
+import lombok.NonNull;
+import lombok.experimental.SuperBuilder;
 
-public class EmailReceiver {
+@SuperBuilder
+public abstract class EmailReceiver {
+    @NonNull String userName;
+    @NonNull String userPassword;
+    @NonNull String serverName;
+    @NonNull String serverPort;
 
-    /**
-     * Test downloading e-mail messages
-     */
-    public static void main(String[] args) {
-        // for IMAP
-        String protocol = "imap";
-        String host = "mail.rpi.edu";
-        String port = "993";
-
-        String userName = "usrName";
-        String password = "password";
-
-        EmailReceiver receiver = new EmailReceiver();
-        receiver.downloadEmails(protocol, host, port, userName, password);
-    }
-
-    /**
-     * Returns a Properties object which is configured for a POP3/IMAP server
-     *
-     * @param protocol either "imap" or "pop3"
-     * @param host
-     * @param port
-     * @return a Properties object
-     */
-    private Properties getServerProperties(String protocol, String host, String port) {
-        Properties properties = new Properties();
-
-        // server setting
-        properties.put(String.format("mail.%s.host", protocol), host);
-        properties.put(String.format("mail.%s.port", protocol), port);
-
-        // SSL setting
-        properties.setProperty(String.format("mail.%s.socketFactory.class", protocol), "javax.net.ssl.SSLSocketFactory");
-        properties.setProperty(String.format("mail.%s.socketFactory.fallback", protocol), "false");
-        properties.setProperty(String.format("mail.%s.socketFactory.port", protocol), String.valueOf(port));
-
-        return properties;
-    }
-
-    /**
-     * Downloads new messages and fetches details for each message.
-     *
-     * @param protocol
-     * @param host
-     * @param port
-     * @param userName
-     * @param password
-     */
-    public void downloadEmails(String protocol, String host, String port, String userName, String password) {
-        Properties properties = getServerProperties(protocol, host, port);
-        Session session = Session.getDefaultInstance(properties);
-
-        try {
-            // connects to the message store
-            Store store = session.getStore(protocol);
-            store.connect(userName, password);
-
-            // opens the inbox folder
-            Folder folderInbox = store.getFolder("INBOX");
-            folderInbox.open(Folder.READ_ONLY);
-
-            // fetches new messages from server
-            Message[] messages = folderInbox.getMessages();
-
-            for (int i = 0; i < messages.length; i++) {
-                Message msg = messages[i];
-                Address[] fromAddress = msg.getFrom();
-                String from = fromAddress[0].toString();
-                String subject = msg.getSubject();
-                String toList = parseAddresses(msg.getRecipients(Message.RecipientType.TO));
-                String ccList = parseAddresses(msg.getRecipients(Message.RecipientType.CC));
-                String sentDate = msg.getSentDate().toString();
-
-                String contentType = msg.getContentType();
-                String messageContent = "";
-
-                if (contentType.contains("text/plain") || contentType.contains("text/html")) {
-                    try {
-                        Object content = msg.getContent();
-                        if (content != null) {
-                            messageContent = content.toString();
-                        }
-                    } catch (Exception ex) {
-                        messageContent = "[Error downloading content]";
-                        ex.printStackTrace();
-                    }
-                }
-
-                // print out details of each message
-                System.out.println("Message #" + (i + 1) + ":");
-                System.out.println("\t From: " + from);
-                System.out.println("\t To: " + toList);
-                System.out.println("\t CC: " + ccList);
-                System.out.println("\t Subject: " + subject);
-                System.out.println("\t Sent Date: " + sentDate);
-                System.out.println("\t Message: " + messageContent);
-            }
-
-            // disconnect
-            folderInbox.close(false);
-            store.close();
-        } catch (NoSuchProviderException ex) {
-            System.out.println("No provider for protocol: " + protocol);
-            ex.printStackTrace();
-        } catch (MessagingException ex) {
-            System.out.println("Could not connect to the message store");
-            ex.printStackTrace();
-        }
-    }
-
-    /**
-     * Returns a list of addresses in String format separated by comma
-     *
-     * @param address an array of Address objects
-     * @return a string represents a list of addresses
-     */
-    private String parseAddresses(Address[] address) {
-        String listAddress = "";
-
-        if (address != null) {
-            for (int i = 0; i < address.length; i++) {
-                listAddress += address[i].toString() + ", ";
-            }
-        }
-        if (listAddress.length() > 1) {
-            listAddress = listAddress.substring(0, listAddress.length() - 2);
-        }
-
-        return listAddress;
-    }
+    public abstract void downloadEmails(String folder);
 }
